Add files via upload
It is a very very draft version of the game "lights off", you can google the name for details.
I want to make it a game on Android, with the size of the grid dynamically changes with the command of the user. However I do not know how to make the gridview change dynamically so that I hard code it to be 3*3. Unfortunately not only my xml file, even the java code I write cannot run because of some errors that I am not able to figure out myself.
Please help me with my stupid java code first. Thank you very much!